+++ HISTORY.md
@@ -11,6 +11,8 @@
 * [RM#16525](https://redmine.iart.ch/issues/16525) Associate Membership - adapt product price, name and make it not updatable
 * [RM#16623](https://redmine.iart.ch/issues/16623) filter industry / expertise: OR not AND condition
 * [RM#16260](https://redmine.iart.ch/issues/16260) Update /my/account design
+* [RM#16622](https://redmine.iart.ch/issues/16622) info message on unpublish
+* [RM#16392](https://redmine.iart.ch/issues/16392) countries pre-sorted
 
 
 ## 9.2.3 (2016-02-14)